Missing Pieces and Pickle support

+++ cspell.py
@@ -127,26 +127,34 @@ class LogParser:
         self.log_cluster_lines = self.parser.parse(self.df_log["Content"])
         t1 = datetime.now()
 
-        logging.info('Parsing done. [Time taken: {!s}]'.format(t0 - t1))
+        logging.info('Parsing done. [Time taken: {!s}]'.format(t1 - t0))
 
         self.cluster_to_df()
 
-        if persistence:
-            self.output_result()
+        try:
 
-        # Update last_line for next execution if called in batch
-        self.set_last_line_id()
+            self.trie_root = self.parser.trieRoot
 
-        root_node_path = os.path.join(self.save_path, 'rootNode.pkl')
-        log_clu_l_path = os.path.join(self.save_path, 'logCluL.pkl')
-        logging.info(f'rootNodePath: {root_node_path}')
-        with open(root_node_path, 'wb') as output:
-            pickle.dump(self.trie_root, output, pickle.HIGHEST_PROTOCOL)
-        logging.info(f'logCluLPath: {log_clu_l_path}')
-        with open(log_clu_l_path, 'wb') as output:
-            pickle.dump(self.log_cluster_lines, output, pickle.HIGHEST_PROTOCOL)
+            if not os.path.exists(self.save_path):
+                os.makedirs(self.save_path)
+            if persistence:
+                self.output_result()
 
-        logging.info('Saving done (persistence={!s}). [Time taken: {!s}]'.format(persistence, datetime.now() - t1))
+            # Update last_line for next execution if called in batch
+            self.set_last_line_id()
+
+            root_node_path = os.path.join(self.save_path, 'rootNode.pkl')
+            log_clu_l_path = os.path.join(self.save_path, 'logCluL.pkl')
+            logging.info(f'rootNodePath: {root_node_path}')
+            with open(root_node_path, 'wb') as output:
+                pickle.dump(self.trie_root, output, pickle.HIGHEST_PROTOCOL)
+            logging.info(f'logCluLPath: {log_clu_l_path}')
+            with open(log_clu_l_path, 'wb') as output:
+                pickle.dump(self.log_cluster_lines, output, pickle.HIGHEST_PROTOCOL)
+
+            logging.info('Saving done (persistence={!s}). [Time taken: {!s}]'.format(persistence, datetime.now() - t1))
+        except Exception as e:
+            return self.df_log
 
         return self.df_log
 
@@ -202,17 +210,23 @@ class LogParser:
         return parameter_list
 
     def output_result(self):
+
         if self.main_log_name:
             main_structured_path = os.path.join(self.save_path, self.main_log_name + '_main_structured.csv')
-            df_log_main_structured = pd.read_csv(main_structured_path)
-            last_main_line_id = df_log_main_structured['LineId'].max()
-            # logging.info(f'last_main_line_id: {last_main_line_id}')
-            trimmed = self.df_log[self.df_log['LineId'] > last_main_line_id]
-            df_log_main_structured = pd.concat([df_log_main_structured, trimmed])
+            if os.path.isfile(main_structured_path):
+                df_log_main_structured = pd.read_csv(main_structured_path)
+                last_main_line_id = df_log_main_structured['LineId'].max()
+                # logging.info(f'last_main_line_id: {last_main_line_id}')
+                trimmed = self.df_log[self.df_log['LineId'] > last_main_line_id]
+                df_log_main_structured = pd.concat([df_log_main_structured, trimmed])
 
-            df_log_main_structured.to_csv(os.path.join(self.save_path, self.main_log_name + '_structured.csv'),
-                                          index=False)
-            self.df_event.to_csv(os.path.join(self.save_path, self.main_log_name + '_main_templates.csv'), index=False)
+                df_log_main_structured.to_csv(os.path.join(self.save_path, self.main_log_name + '_structured.csv'),
+                                              index=False)
+                self.df_event.to_csv(os.path.join(self.save_path, self.main_log_name + '_main_templates.csv'),
+                                     index=False)
+            else:
+                self.df_log.to_csv(os.path.join(self.save_path, self.main_log_name + '_structured.csv'), index=False)
+                self.df_event.to_csv(os.path.join(self.save_path, self.main_log_name + '_templates.csv'), index=False)
 
         self.df_log.to_csv(os.path.join(self.save_path, self.log_name + '_structured.csv'), index=False)
         self.df_event.to_csv(os.path.join(self.save_path, self.log_name + '_templates.csv'), index=False)
